+++ python/paddle/fluid/tests/unittests/test_lbfgs.py
@@ -108,10 +108,10 @@ class TestLbfgs(unittest.TestCase):
             # All these minima may be reached from appropriate starting points.
             return 3 * x**4 + 0.4 * x**3 - 5.64 * x**2 + 2.112 * x
 
-        x0 = np.array([0.82], dtype='float32')
+        x0 = np.array([0.82], dtype='float64')
         x1 = np.array([-1.3], dtype='float32')
 
-        results = test_static_graph(func, x0, dtype='float32')
+        results = test_static_graph(func, x0, dtype='float64')
         self.assertTrue(np.allclose(0.8, results[2]))
 
         results = test_dynamic_graph(func, x1, dtype='float32')
@@ -149,14 +149,15 @@ class TestLbfgs(unittest.TestCase):
 
         x0 = np.random.random(size=[2]).astype('float32')
         H0 = np.array([[1.0, 0.0], [0.0, 1.0]]).astype('float32')
-        H1 = np.array([[1.0, 2.0], [2.0, 1.0]]).astype('float32')
+        H1 = np.array([[1.0, 2.0], [2.0, 1.0]]).astype('float64')
         H2 = np.array([[1.0, 2.0], [2.0, 1.0]]).astype('float32')
 
         results = test_static_graph(func, x0, H0)
         self.assertTrue(np.allclose([0., 0.], results[2]))
         self.assertTrue(results[0][0])
 
-        self.assertRaises(ValueError, test_dynamic_graph, func, x0, H0=H1)
+        self.assertRaises(
+            ValueError, test_dynamic_graph, func, x0, H0=H1, dtype='float64')
         self.assertRaises(ValueError, test_dynamic_graph, func, x0, H0=H2)
 
     def test_static_line_search_fn(self):
